fix : 국가 셀렉트 박스 선택 구현 완료

+++ src/main/java/com/encore/bbs/board/dto/BbsDTO.java
@@ -21,6 +21,6 @@ public class BbsDTO {
 	private int likeCnt;
 	private int hashtagId;
 	private boolean isDeleted;
-	private Integer countryId;
+	private Long countryId;
 	private String memberId;
 }
